Properly handle multi channel commands such as NICK and QUIT
Also ignore the multi channel AWAY message while at it. Contrary to QUIT
and NICK it occurs way to often and is not as essential as the former to
messages. Additionally, other IRC clients such as ii or irssi also seem
to ignore it.

This also requires an update of the girc submodule.

See: https://github.com/lrstanley/girc/pull/15