hii: Make indent also default to the user's login name
The data stored in the GECOS field is very likely not a valid ident as
it may contain white spaces. This is not to say that usernames might not
also be bad user/ident values. However, I believe that this is less
likely and thus assume this to be the more sensible default.