Record mention in a separate file
This rather difficult to implement in the frontend and was thus
implemented in the backend. The only potential problem with the code
introduced here is that the regex isn't configurable. However, I believe
that the hardcoded regex will fit most people and those who don't like
it just need to compile there own version of hii.