Use MONITOR command for private conversations
This is useful in order to create a listener for them as soon as they go
online and remove that listener as soon as they go offline again.

This requires support for the IRCv3.2 MONITOR command [0] on the server
side. If the server doesn't support the command the listeners for
private conversations are not removed/recreated automatically which
isn't a big deal.

Additionally, changes were made to the createListener() and
removeListener() function. These function now return a variable error
when removing a listener which doesn't exists or attempting to create a
listener which already exists.

[0]: https://ircv3.net/specs/core/monitor-3.2.html