getSourceDirs: Map to private conversation directories as well
If a user X started a private conversation with user Y and user Y
changes his nick to Z the nick change will be mapped to the directory
for the private conversation as well. Same with quit messages.

Previously, these messages were only mapped to the channels user X and Y
shared.