Add support for automatically monitoring users on startup
This is nice if you constantly have a private query with a user on the
server. Without this patch the listener for this conversation was
removed if hii(1) was restarted, e.g. as a result of a PING timeout.
Therefore requiring the users to manually create a new listener for the
conversation e.g. using:

	echo "/PRIVMSG <user> :<text>" > ~/irc/server/in

This is now no longer necessary if the nick of this user was specified
in the target arguments.