removeListener: don't open FIFO with O_NONBLOCK
O_NONBLOCK was added to the openFifo flags in
053dca083f9e1c1e425584c80b5ff012a93c61bc in order to avoid deadlocks.
Unfortunately, this commit introduced a potential race condition and
depends on behaviour not specified in POSIX.1-2008 (opening a FIFO with
O_NONBLOCK).

Technically speaking removeListener() should never be called if there is
no reader on the FIFO as createListener() doesn't return (without an
error) unless the recvInput goroutine has been started. Additionally,
the recvInput() goroutine itself doesn't terminate. Therefore there
should never be a deadlock in removeListener().