openFifo: remove helper function
Instead use os.OpenFile (which doesn't check the file type though). This
approach has the advantage that we can handle an mkfifo error in
createListener (where it is possible to terminate the program).

The worst case problem resulting of the fact that the file type is no
longer checked is if the file is removed between the syscall.Mkfifo in
createListener() and the os.Open() in recvInput(). In this case
recvInpu() would constantly require CPU time as os.Open() no longer
blocks. But this is (a) very unlikely and (b) requires the user to
manually remove the file.