GNUmakefile: add dist target
We need a dist target because the releases automatically generated by
GitHub for tag names do not contain the source code for the git
submodule.

The name of the tarball and the name of the root directory contained in
it should contain the version number. Unfortunately, figuring out the
current version would require hard coding it or invoking git-describe
from the Makefile. The latter would introduce a dependency on git for
building the software while the former would require updating the hard
coded version with each new tag. Both seem annoying therefore the
version has to be set manually, e.g. `VERSION=4.2 make dist`.

Discussion: In the long run we probably want to hard code the current
version in the Makefile and add a `-v` flag to hii which outputs that
hard coded version.