hii: fsync on directory containing the renamed file
This is required to ensure that the file is actually written to disc.
Without this this change we might end up with a zero-length file.

Discussion: An alternative to implementing atomic file writes using
rename(3) own our own is simply using renameio [0]. However, I am unsure
if introducing a new dependency is really necessary.

[0]: https://github.com/google/renameio