removeListener: resolve data races regarding listener access
Previously the listener was created by the serveNicks() goroutine in the
background. The goroutine might not have been started before
removeListener() is called.

This also means that the `dir.ch.ln != nil` check in removeListener is
no longer required because if the listener was created it is guaranteed
that the listener exists.

This issue was found using the go race detector.