getSourceDirs: Don't return an error
The return value of Client.LookupUser should never be nil as we only
call this function for existing users. If it is indeed nil this is a
programming error not a runtime error and should be fixed. We could but
an explicit panic there but I don't think that this is necessary.

This allows us to shrink the code a bit and remove the die function.