fmtEvent: Remove non-printable characters
Most importantly, this disables any ANSI escape sequences with which
various nasty things could potentially be done [0]. One might argue that
non-printable characters could also be filtered in the frontend,
however, I believe that it is safer to do it in the backend.

Since all non-printable characters are removed now it should no longer
be necessary to call girc.StripRaw to remove color specific control
bytes.

The debug output is not sanitized though, this has been documented.

[0]: https://www.openwall.com/lists/oss-security/2015/08/11/8