writeEvent: Add non-printable character suffixes to some lines
This makes it possible to detect messages mentioning the own users and
messages written by the own users. Compared to exporting the user name
using an atomically written file in the backend and doing this
distinction in the frontend this solution has the advantage that it also
works for older message written at a time where the user used a nick
different from the one he uses currently.

Messages written by the own user are suffixed with an ACK character,
messages mentioning the users are suffixed with a BEL character. These
suffixes are inserted before the newline character.

See also: https://github.com/nmeum/hii/issues/1