fmtEvent: Call girc.StripRaw again
The call was removed in c5077da376cb13d3ad0374e63bb272d333d4d307
because the change introduced in this commit removes all escape
sequences. However, the girc.StripRaw function has the slight advantage
that it also removes the color identification code.

Not stripping the color code is, for instance, slightly annoying in IRC
channels with bots that always use color codes.