Increase default PING delay to one minute
By default girc sends a PING command every 20 seconds which seems
a bit excessive. Irssi sends a PING every 60 seconds which seems
like a more reasonable default.

Irssi also the configuration of a separate timespan for
disconnecting/reconnecting if no PONG is received in a certain interval.
Unfortunately, this can presently not be configured for girc.

See also:

* https://github.com/irssi/irssi/blob/42f83fc130caa4349646ea967ffafa4de4006024/docs/manual.txt#L442-L473
* https://github.com/lrstanley/girc/pull/57